requirements.txt
numpy
pandas
matplotlib
pillow
scikit-learn
torch
torchvision
notebook
kaggle
streamlit